+++ lib/ezy_form.dart
@@ -7,9 +7,12 @@
 library ezy_form;
 
 export 'src/form_array_widget.dart';
+export 'src/form_control_watcher.dart';
 export 'src/form_control_widget.dart';
+export 'src/form_watcher.dart';
 export 'src/form_group_widget.dart';
 export 'src/models/form_array_control.dart';
 export 'src/models/form_control.dart';
 export 'src/models/form_group.dart';
+export 'src/models/form_node.dart';
 export 'src/validators.dart';